// think-big-2026/backend/src/services/certificateService.d.ts
export declare const generateCertificate: (data: {
    teamName: string;
    studentName: string;
    registerNumber: string;
    projectTitle: string;
    eventName: string;
    certificateType: string;
    certificateId: string;
    department: string;
    year?: string;
}) => Promise<string>;
//# sourceMappingURL=certificateService.d.ts.map